submit java-collections-ru/streams

+++ java-collections-ru/streams/src/main/java/exercise/App.java
@@ -5,7 +5,7 @@ import java.util.Arrays;
 
 public class App {
     public static void getCountOfFreeEmails(List<String> emailList) {
-        int countFreeEmails = emailList.stream()
+        long countFreeEmails = emailList.stream()
                 .filter(email -> email.contains("gmail.com") || email.contains("hotmail.com") || email.contains("yandex.ru"))
                 .count();
         System.out.println(countFreeEmails);
